refactored inference engines
