+++ docs/source/examples/replicate_vicuna/translate_opts_ct2.yaml
@@ -6,7 +6,7 @@ tgt_subword_model: llama/tokenizer.model
 src_subword_vocab: finetuned_llama7B/llama7B-vicuna-onmt_step_4000.concat_CT2/vocabulary.json
 world_size: 1
 gpu_ranks: [0]
-max_length: 50
+max_length: 512
 batch_type: sents
 batch_size: 1
 precision: fp16
@@ -14,4 +14,6 @@ verbose: true
 random_sampling_topk: 40
 random_sampling_topp: 0.75
 random_sampling_temp: 0.1
+beam_size: 3
+n_best: 2
 src: None
