+++ docs/source/examples/replicate_vicuna/translate_opts_py.yaml
@@ -3,8 +3,12 @@ model: finetuned_llama7B/llama7B-vicuna-onmt_step_4000.concat.pt
 transforms: [sentencepiece]
 src_subword_model: llama/tokenizer.model
 tgt_subword_model: llama/tokenizer.model
+# world_size: 2
+# gpu_ranks: [0, 1]
+# parallel_mode: tensor_parallel
 world_size: 1
 gpu_ranks: [0]
+gpu: 0
 max_length: 512
 batch_type: sents
 batch_size: 1
